add app.py and transform_logs()
Add the steamlit app with a text box for the userToken.
Start work on transform_logs(). The goal is to have a usable NxN
Dataframe without inner objects / lists. One row should correspond to
the data of one player in one fight. This will add some redundancy as
the metadata of each log is stored for every player (~15 times) but that
seams to be managable.