add multithreaded fetch
Create a fixed number of worker threads, that will take log_ids from a
queue and then merge the logs afterwards.

In order for this to work the streamlit.spinner, that is usually
displayed for cached function calls needs to be disabled because
there is a bug in the streamlit ui that sometimes throws an error
when too many need to be displayed.